Armada folder missing inside /opt/platform
The armada folder do not exist inside the /opt/platform
when the host do not have any armada application
raising an error when armada migration config is
triggered on 22.06 -> 22.12 upgrade

Test Plan:

PASS:
CentOS -> CentOS DX upgrade:
- 22.06 -> 22.12

Story: 2009303
Task: 45952

+++ controllerconfig/controllerconfig/controllerconfig/upgrades/controller.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2016-2020 Wind River Systems, Inc.
+# Copyright (c) 2016-2022 Wind River Systems, Inc.
 #
 # SPDX-License-Identifier: Apache-2.0
 #
@@ -306,6 +306,11 @@ def migrate_pxeboot_config(from_release, to_release):
 
 def migrate_armada_config(from_release, to_release):
     """ Migrates armada configuration. """
+
+    # Check if the folder exist before migration
+    if not os.path.exists(os.path.join(PLATFORM_PATH, "armada")):
+        LOG.info("Skipping armada migration, the directory doesn't exist")
+        return
 
     LOG.info("Migrating armada config")
     devnull = open(os.devnull, 'w')
